fix: rename base input data class
