+++ src/use_case/theoretical_time/base_input_data.ts
